requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
mysql-connector-python>=8.0.0

# Security
bcrypt>=4.0.0
python-dotenv>=1.0.0

# PDF generation
reportlab>=4.0.0

# Data processing
numpy>=1.24.0

# Optional but recommended
Pillow>=10.0.0
requests>=2.31.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0